move types to separate folder
